+++ back/src/main/java/com/ll/goohaeyou/jobPost/jobPost/application/JobPostEventListener.java
@@ -1,0 +1,34 @@
+package com.ll.goohaeyou.jobPost.jobPost.application;
+
+import com.ll.goohaeyou.global.event.notification.PostDeadlineEvent;
+import com.ll.goohaeyou.global.event.notification.PostEmployedEvent;
+import com.ll.goohaeyou.jobPost.jobPost.domain.JobPost;
+import com.ll.goohaeyou.jobPost.jobPost.domain.repository.JobPostRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.event.EventListener;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
+
+@Component
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class JobPostEventListener {
+
+    private final JobPostRepository jobPostRepository;
+
+    @EventListener
+    @Transactional
+    public void jobPostClosedEventListen(PostDeadlineEvent event) {
+        JobPost jobPost = event.getJobPost();
+        jobPost.close();
+        jobPostRepository.save(jobPost);
+    }
+
+    @EventListener
+    @Transactional
+    public void jobPostEmployedEventListen(PostEmployedEvent event) {
+        JobPost jobPost = event.getJobPost();
+        jobPost.employed();
+        jobPostRepository.save(jobPost);
+    }
+}
